pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "weftlyflow"
version = "0.1.0a0"
description = "Self-hosted workflow automation platform — Python backend."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Weftlyflow contributors" }]
keywords = ["workflow", "automation", "orchestration", "webhooks", "ai-agents"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
]

dependencies = [
    # web
    "fastapi>=0.110",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    # data
    "pydantic[email]>=2.6",
    "pydantic-settings>=2.2",
    "sqlalchemy>=2.0.29",
    "alembic>=1.13",
    "psycopg[binary]>=3.1",
    "aiosqlite>=0.19",
    # queue + schedule
    "celery[redis]>=5.3",
    "redis>=5.0",
    "apscheduler>=3.10",
    # security / crypto / auth
    "cryptography>=42.0",
    "pyjwt>=2.8",
    "argon2-cffi>=23.1",
    "authlib>=1.3",
    "pyotp>=2.9",
    # sandbox
    "RestrictedPython>=7.1",
    # observability
    "structlog>=24.1",
    "prometheus-client>=0.20",
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    # cli + misc
    "typer>=0.12",
    "python-ulid>=2.2",
    "orjson>=3.10",
    "anyio>=4.3",
    "jinja2>=3.1",      # internal templates only (emails), NOT the expression engine
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-xdist>=3.5",
    "hypothesis>=6.99",
    "respx>=0.21",
    "fakeredis>=2.21",
    "freezegun>=1.4",
    "ruff>=0.4",
    "black>=24.4",
    "isort>=5.13",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "types-python-dateutil",
]
docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "mkdocs-gen-files>=0.5",
    "mkdocs-literate-nav>=0.6",
    "mkdocs-section-index>=0.3",
]
ai = [
    "langchain>=0.2",
    "langchain-community>=0.2",
    "openai>=1.30",
    "anthropic>=0.28",
    "tiktoken>=0.7",
]
sso = [
    "python3-saml>=1.16",
]
all = ["weftlyflow[dev,docs,ai,sso]"]

[project.scripts]
weftlyflow = "weftlyflow.cli:app"

[project.entry-points."weftlyflow.nodes"]
# Third-party nodes can register themselves here. Built-ins are auto-discovered.

[project.urls]
Homepage = "https://example.com/weftlyflow"
Documentation = "https://example.com/weftlyflow/docs"
Source = "https://example.com/weftlyflow/source"

# ----------------------------------------------------------------------------
# Tooling
# ----------------------------------------------------------------------------

[tool.hatch.build.targets.wheel]
packages = ["src/weftlyflow"]

[tool.hatch.build.targets.sdist]
include = ["/src", "/tests", "/README.md", "/LICENSE", "/IMPLEMENTATION_BIBLE.md"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "B",        # bugbear
    "C4",       # comprehensions
    "UP",       # pyupgrade
    "N",        # naming
    "SIM",      # simplify
    "RUF",      # ruff-specific
    "PL",       # pylint
    "PTH",      # pathlib
    "ERA",      # no-commented-out-code
    "D",        # pydocstyle (Google)
]
ignore = [
    "D203",   # incompatible with D211
    "D213",   # incompatible with D212
    "PLR0913", # too-many-args — fine in our domain
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D", "PLR2004", "S101", "PLC0415"]
"src/weftlyflow/db/migrations/versions/*.py" = ["D", "E501"]
"scripts/**" = ["D"]
# FastAPI's standard `Depends(...)` / `Query(...)` / `Header(...)` argument
# defaults trip B008; they are the idiomatic pattern and safe because FastAPI
# re-evaluates each call per request.
"src/weftlyflow/server/deps.py" = ["B008"]
"src/weftlyflow/server/routers/**" = ["B008"]
"src/weftlyflow/server/middleware.py" = ["BLE001"]
"src/weftlyflow/server/routers/health.py" = ["BLE001"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["weftlyflow"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["RestrictedPython.*", "apscheduler.*", "authlib.*", "celery.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
asyncio_mode = "auto"
markers = [
    "unit: pure unit test (default)",
    "integration: touches DB/Redis/FS",
    "node: per-node test",
    "live: hits real external services (opt-in)",
    "load: performance / load test",
]

[tool.coverage.run]
source = ["src/weftlyflow"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "\\.\\.\\.$",
]
fail_under = 80
